pyproject.toml
[project]
name = "hello-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "google-genai>=1.33.0",
    "mcp[cli]>=1.13.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
